refactor: author-decorator to auth-module
