+++ src/discord/discord.controller.ts
@@ -11,7 +11,7 @@ import {
 } from '@nestjs/common';
 import { ApiHeader, ApiSecurity, ApiTags } from '@nestjs/swagger';
 import { TokenGuard } from '../auth/token.strategy';
-import { Author, AuthorObject } from '../util/getAuthorFromHeaders.decorator';
+import { Author, AuthorObject } from '../auth/getAuthorFromHeaders.decorator';
 import { DiscordService } from './discord.service';
 import { DiscordDTO } from './dto/discord.dto';
 @ApiTags('Discord')
