+++ src/discord/interfaces/discord.interface.ts
@@ -1,4 +1,4 @@
-import { Author } from '../../util/getAuthorFromHeaders.decorator';
+import { Author } from '../../auth/getAuthorFromHeaders.decorator';
 
 export interface DiscordProfile {
   author: DiscordProfileAuthor;
